chore: deserialize formatter options
