+++ crates/rome_js_formatter/Cargo.toml
@@ -16,11 +16,13 @@ categories = [
 [dependencies]
 cfg-if = "1.0.0"
 rome_js_syntax = { version = "0.0.2", path = "../rome_js_syntax" }
+rome_json_syntax = { version = "0.0.1", path = "../rome_json_syntax" }
 rome_js_factory = { version = "0.0.2", path = "../rome_js_factory" }
 rome_formatter = { version = "0.0.1", path = "../rome_formatter" }
 rome_rowan = { version = "0.0.1", path = "../rome_rowan" }
 rome_text_size = { version = "0.0.1", path = "../rome_text_size" }
 rome_diagnostics_categories = { version = "0.0.1", path = "../rome_diagnostics_categories" }
+rome_deserialize = { path = "../rome_deserialize", version = "0.0.0" }
 tracing = { workspace = true }
 unicode-width = "0.1.9"
 serde = { version = "1.0.136", features = ["derive"], optional = true }
